Account

+++ app/Http/Controllers/HRM/DepartmentController.php
@@ -18,10 +18,7 @@ class DepartmentController extends Controller
 
     public function index(Request $request)
     {
-        if (!auth()->user()->can('departments_index')) {
-
-            abort(403, 'Access Forbidden.');
-        }
+        abort_if(!auth()->user()->can('departments_index') || config('generalSettings')['subscription']->features['hrm'] == 0, 403);
 
         if ($request->ajax()) {
 
@@ -33,20 +30,14 @@ class DepartmentController extends Controller
 
     public function create()
     {
-        if (!auth()->user()->can('departments_create')) {
-
-            abort(403, 'Access Forbidden.');
-        }
+        abort_if(!auth()->user()->can('departments_create') || config('generalSettings')['subscription']->features['hrm'] == 0, 403);
 
         return view('hrm.departments.ajax_view.create');
     }
 
     public function store(Request $request)
     {
-        if (!auth()->user()->can('departments_create')) {
-
-            abort(403, 'Access Forbidden.');
-        }
+        abort_if(!auth()->user()->can('departments_create') || config('generalSettings')['subscription']->features['hrm'] == 0, 403);
 
         $this->departmentService->storeValidation(request: $request);
         return $this->departmentService->addDepartment(request: $request);
@@ -54,10 +45,7 @@ class DepartmentController extends Controller
 
     public function edit($id)
     {
-        if (!auth()->user()->can('departments_edit')) {
-
-            abort(403, 'Access Forbidden.');
-        }
+        abort_if(!auth()->user()->can('departments_edit') || config('generalSettings')['subscription']->features['hrm'] == 0, 403);
 
         $department = $this->departmentService->singleDepartment(id: $id);
 
@@ -66,10 +54,7 @@ class DepartmentController extends Controller
 
     public function update($id, Request $request)
     {
-        if (!auth()->user()->can('departments_edit')) {
-
-            abort(403, 'Access Forbidden.');
-        }
+        abort_if(!auth()->user()->can('departments_edit') || config('generalSettings')['subscription']->features['hrm'] == 0, 403);
 
         $this->departmentService->updateValidation(request: $request, id: $id);
         $this->departmentService->updateDepartment(request: $request, id: $id);
@@ -79,10 +64,7 @@ class DepartmentController extends Controller
 
     public function delete($id)
     {
-        if (!auth()->user()->can('departments_delete')) {
-
-            abort(403, 'Access Forbidden.');
-        }
+        abort_if(!auth()->user()->can('departments_delete') || config('generalSettings')['subscription']->features['hrm'] == 0, 403);
 
         $this->departmentService->deleteDepartment(id: $id);
